Move logging interceptor tests to a separate file
